+++ tests/testthat/test-outbreak.R
@@ -1,20 +1,20 @@
 context("Outbreak")
 library(dfoliatR)
 
-data("ef_obr")
+data("efk_obr")
 
-test_obr <- outbreak(dfoliatR::ef_defol)
+test_obr <- outbreak(efk_defol)
 
 obr_stats <- outbreak_stats(test_obr)
 
 test_that("outbreak consistently defines outbreaks", {
-  expect_equal(ef_obr$outbreak_status, test_obr$outbreak_status)
+  expect_equal(efk_obr$outbreak_status, test_obr$outbreak_status)
 })
 
 test_that("outbreak_stats gets the same first interval", {
-  expect_equivalent(data.frame(1790, 1808), obr_stats[1, 1:2])
+  expect_equivalent(data.frame(1788, 1796), obr_stats[1, 1:2])
 })
 
 test_that("outbreak stats finds the same number of events", {
-  expect_equal(nrow(obr_stats), 9)
+  expect_equal(nrow(obr_stats), 10)
 })